requirements.txt
transformers
pillow
torch
protobuf
ipywidgets
tqdm
tokenizers
sentencepiece
chromadb
fastapi
uvicorn[standard]
jinja2
google-generativeai
sentence-transformers
scikit-learn
langchain
pydantic
python-dotenv